+++ src/components/Menu/Menu.tsx
@@ -1,7 +1,7 @@
 import { motion, Variants } from 'framer-motion';
 import * as React from 'react';
 import styled from 'styled-components';
-import { PopperProps, useDisclosure, usePopper } from '../../hooks';
+import { PopperProps, useDisclosure, UseDisclosureReturn, usePopper } from '../../hooks';
 import { sxMixin, SxProp, useComponentStyles } from '../../system';
 import { AsProp, ChildrenProp } from '../../types';
 import { runIfFn } from '../../util';
@@ -23,9 +23,10 @@ const variants: Variants = {
 };
 
 export interface MenuProps extends PopperProps, AsProp, ChildrenProp, SxProp {
-  trigger: ((props: { isOpen: boolean }) => React.ReactElement) | React.ReactElement;
-  children: ((props: { isOpen: boolean }) => React.ReactElement) | React.ReactNode;
+  trigger: ((props: UseDisclosureReturn) => React.ReactElement) | React.ReactElement;
+  children: ((props: UseDisclosureReturn) => React.ReactElement) | React.ReactNode;
   isFullWidth?: boolean;
+  openOnHover?: boolean;
 }
 
 export const Menu: React.FC<MenuProps> = (props) => {
@@ -37,6 +38,7 @@ export const Menu: React.FC<MenuProps> = (props) => {
     closeOnEsc,
     closeOnInnerClick,
     closeOnOuterClick,
+    openOnHover,
     placement,
     offset,
     ...rest
@@ -44,7 +46,7 @@ export const Menu: React.FC<MenuProps> = (props) => {
 
   const styles = useComponentStyles('Menu', props);
 
-  const { isOpen, onClose, onToggle } = useDisclosure();
+  const { isOpen, onOpen, onClose, onToggle } = useDisclosure();
 
   const { triggerProps, childProps } = usePopper({
     isOpen,
@@ -59,10 +61,11 @@ export const Menu: React.FC<MenuProps> = (props) => {
 
   return (
     <>
-      {React.cloneElement(runIfFn(trigger, { isOpen }), {
+      {React.cloneElement(runIfFn(trigger, { isOpen, onOpen, onClose, onToggle }), {
         ...triggerProps,
-        onClick: onToggle,
+        onClick: openOnHover ? undefined : onToggle,
         onKeyDown: ({ key }: React.KeyboardEvent<HTMLDivElement>) => (key === 'Enter' ? onToggle : undefined),
+        onMouseEnter: openOnHover ? onOpen : undefined,
       })}
       <Portal>
         <Motion
@@ -74,7 +77,7 @@ export const Menu: React.FC<MenuProps> = (props) => {
           sx={styles}
           {...rest}
         >
-          {runIfFn(children, { isOpen })}
+          {runIfFn(children, { isOpen, onOpen, onClose, onToggle })}
         </Motion>
       </Portal>
     </>
@@ -86,6 +89,7 @@ Menu.defaultProps = {
   closeOnEsc: true,
   closeOnInnerClick: true,
   closeOnOuterClick: true,
+  openOnHover: false,
   placement: 'bottom-start',
   offset: [0, 8],
 };