+++ src/components/Theme/injectGlobalStyles.tsx
@@ -115,6 +115,10 @@ export const injectGlobalStyles = (additionalStyles?: string | ReturnType<typeof
 
     /* ${animationKeyframes} */
 
+    .noscroll {
+      overflow: hidden;
+    }
+
     ${additionalStyles};
   `;